Use setwise bitboard operations for evaluation

+++ engine/eval_utils_test.go
@@ -89,13 +89,13 @@ func TestBackwardPawns(t *testing.T) {
 	fen := "rnbqkbnr/5ppp/p2p4/P2P1P2/1P2P3/8/6PP/RNBQKBNR w KQkq - 0 1"
 	game := FromFen(fen, true)
 
-	expected := 2
+	expected := int16(2)
 	actual := game.Position().CountBackwardPawns(White)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
 	}
 
-	expected = 2
+	expected = int16(2)
 	actual = game.Position().CountBackwardPawns(Black)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
@@ -106,13 +106,13 @@ func TestCandidatePawns(t *testing.T) {
 	fen := "7k/p7/8/PP6/5ppp/8/5P1P/K7 w - - 0 1"
 	game := FromFen(fen, true)
 
-	expected := 1
+	expected := int16(1)
 	actual := game.Position().CountCandidatePawns(White)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
 	}
 
-	expected = 1
+	expected = int16(1)
 	actual = game.Position().CountCandidatePawns(Black)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
@@ -123,13 +123,13 @@ func TestDoublePawns(t *testing.T) {
 	fen := "7k/1p6/8/PP3p1p/P4p1p/PP6/8/K7 w - - 0 1"
 	game := FromFen(fen, true)
 
-	expected := 3
+	expected := int16(3)
 	actual := game.Position().CountDoublePawns(White)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
 	}
 
-	expected = 2
+	expected = int16(2)
 	actual = game.Position().CountDoublePawns(Black)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
@@ -140,13 +140,13 @@ func TestIsolatedPawns(t *testing.T) {
 	fen := "7k/8/2p1P3/P2p1p1p/5p2/6P1/7P/K7 w - - 0 1"
 	game := FromFen(fen, true)
 
-	expected := 2
+	expected := int16(2)
 	actual := game.Position().CountIsolatedPawns(White)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
 	}
 
-	expected = 3
+	expected = int16(3)
 	actual = game.Position().CountIsolatedPawns(Black)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
@@ -157,13 +157,13 @@ func TestPassedPawns(t *testing.T) {
 	fen := "7k/8/2p1P3/P2p1p1p/5p2/6P1/7P/K7 w - - 0 1"
 	game := FromFen(fen, true)
 
-	expected := 2
+	expected := int16(2)
 	actual := game.Position().CountPassedPawns(White)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
 	}
 
-	expected = 2
+	expected = int16(2)
 	actual = game.Position().CountPassedPawns(Black)
 	if actual != expected {
 		t.Error(fmt.Sprintf("Expected: %d\n, Got: %d\n", expected, actual))
